ai_preprocess/requirements.txt
torch>=2.0.0
transformers>=4.30.0
kafka-python>=2.0.2
